% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chess}
\alias{chess}
\title{Chess Dataset}
\format{A data frame with 2000 rows and 2 variables:
\describe{
  \item{x1}{random generated}
  \item{x2}{random generated}
}}
\usage{
chess
}
\description{
Bidimensional binary dataset to be used in toy examples.
}
\keyword{datasets}

