% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FastKNN.R
\name{Distance_for_KNN_test}
\alias{Distance_for_KNN_test}
\title{Distance for KNN Test
The Distance_for_KNN_test returns the distance matrix between the test set and the training set.}
\usage{
Distance_for_KNN_test(test_set, train_set)
}
\arguments{
\item{test_set}{is a matrix where the columns are the features of the test set}

\item{train_set}{is a matrix with the features of the training set}
}
\value{
a distance matrix
}
\description{
Distance for KNN Test
The Distance_for_KNN_test returns the distance matrix between the test set and the training set.
}
\seealso{
\code{knn_test_function}

\code{pdist}
}

