% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastKNN-package.R
\docType{package}
\name{FastKNN-package}
\alias{FastKNN}
\alias{FastKNN-package}
\title{FastKNN: Fast Nearest Neighbour Search (Wraps ANN Library) Using L2 Metric}
\description{
Finds the k nearest neighbours for a point using Arya and Mount's ANN library (v1.1.3).
}
\author{
\strong{Maintainer}: Christof Schötz \email{math@christof-schoetz.de} (\href{https://orcid.org/0000-0003-3528-4544}{ORCID})

}
\keyword{internal}
