% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{fastknn}
\alias{fastknn}
\title{Fast KNN Classifier}
\usage{
fastknn(xtr, ytr, xte, k, method = "dist")
}
\arguments{
\item{xtr}{matrix containing the training instances.}

\item{ytr}{factor array with the training labels.}

\item{xte}{matrix containing the test instances.}

\item{k}{number of neighbors considered.}

\item{method}{method used to infer the class membership probabilities of the 
test instances. Choose \code{"dist"} (default) to compute probabilites from 
the inverse of the nearest neighbor distances. This method works as
a shrinkage estimator and provides a better predictive performance in general.
Or you can choose \code{"vote"} to compute probabilities from the frequency 
of the nearest neighbor labels.}
}
\value{
\code{list} with predictions for the test set:
\itemize{
 \item \code{class}: factor array of predicted classes.
 \item \code{prob}: matrix with predicted probabilities.
}
}
\description{
Fast k-Nearest Neighbor classifier build upon ANN, a high efficient 
\code{C++} library for nearest neighbor searching.
}

