% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FastKNN.R
\name{k.nearest.neighbors}
\alias{k.nearest.neighbors}
\title{k-Nearest Neighbors
the \code{k.nearest.neigbors} gives the list of points (k-Neigbours) that are closest
to the row i in descending order.}
\usage{
k.nearest.neighbors(i, distance_matrix, k = 5)
}
\arguments{
\item{i}{is from the numeric class and is a row from the distance_matrix.}

\item{distance_matrix}{is a nxn matrix.}

\item{k}{is from the numeric class and represent the number of neigbours that the function will return.}
}
\value{
a k vector with the k closest neigbours to the i observation.
}
\description{
k-Nearest Neighbors
the \code{k.nearest.neigbors} gives the list of points (k-Neigbours) that are closest
to the row i in descending order.
}
\details{
The output of this function is used in the \code{knn_test_function} function.
}
\seealso{
\code{order}
}

