% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/FastKNN.R
\name{knn_training_function}
\alias{knn_training_function}
\title{KNN Training
The knn_training_function returns the labels for a training set using the
k-Nearest Neighbors Clasification method.}
\usage{
knn_training_function(dataset, distance, label, k = 1)
}
\arguments{
\item{dataset}{is a matrix with the features of the training set}

\item{distance}{is a nxn matrix with the distance between each observation of the training set}

\item{label}{is a nx1 vector with the labels of the training set}

\item{k}{is from the numeric class and represent the number of neigbours to be use in the classifier.}
}
\value{
a k vector with the predicted labels for the training set.
#'
}
\description{
KNN Training
The knn_training_function returns the labels for a training set using the
k-Nearest Neighbors Clasification method.
}
\details{
This function is use to check the quality of the Classifier. Because then the predicted labels
are compared with the true labels
}
\seealso{
\code{k.nearest.neighbors}

\code{sample}
}

