% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{knnDecision}
\alias{knnDecision}
\title{Classification Decision Boundary}
\usage{
knnDecision(xtr, ytr, xte, yte, k, method = "dist", dpi = 150,
  show = TRUE)
}
\arguments{
\item{xtr}{matrix containing the training instances. If \code{xtr} is not a 
bidimensional matrix only the first two columns will be considered.}

\item{ytr}{factor array with the training labels.}

\item{xte}{(optional) Matrix containing the test instances. The test points will be 
plotted over the surface boundary. If missing, the training points will be 
plotted instead. If \code{xte} is not a bidimensional matrix only the first 
two columns will be considered.}

\item{yte}{(optional) Factor array with the test labels.}

\item{k}{number of neighbors considered.}

\item{method}{method used to infer the class membership probabilities of the 
test instances. Choose \code{"dist"} (default) to compute probabilites from 
the inverse of the nearest neighbor distances. This method works as
a shrinkage estimator and provides a better predictive performance in general.
Or you can choose \code{"vote"} to compute probabilities from the frequency 
of the nearest neighbor labels.}

\item{dpi}{a scalar that defines the graph resolution (default = 150). 
It means that \code{dpi^2} data points will be generated from the original 
dataset to draw the decision boundary. So, for large values (>= 300) it may 
take too much time to plot.}

\item{show}{logical: if TRUE (the default option), the generated \code{ggplot2}
object will be plotted.}
}
\value{
\code{ggplot2} object.
}
\description{
Decision boundary of the \code{fastknn} classifier.
}

