% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multi_spirals}
\alias{multi_spirals}
\title{Multi Spirals Dataset}
\format{A data.frame with 2000 rows and 3 variables:
\describe{
  \item{x1}{randomly generated numeric feature}
  \item{x2}{randomly generated numeric feature}
  \item{target}{factor array with class labels in 4 levels}
}}
\usage{
multi_spirals
}
\description{
Bidimensional multiclass dataset to be used as a toy example.
}
\seealso{
Other Dataset: \code{\link{chess}},
  \code{\link{covertype}}, \code{\link{spirals}}
}
\keyword{datasets}

