% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spirals}
\alias{spirals}
\title{Spirals Dataset}
\format{A data.frame with 2000 rows and 3 variables:
\describe{
  \item{x1}{random generated}
  \item{x2}{random generated}
  \item{target}{class labels}
}}
\usage{
spirals
}
\description{
Bidimensional binary dataset to be used in toy examples.
}
\keyword{datasets}

