% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chess}
\alias{chess}
\title{Chess Dataset}
\format{A data.frame with 2000 rows and 3 variables:
\describe{
  \item{x1}{randomly generated numeric feature}
  \item{x2}{randomly generated numeric feature}
  \item{target}{factor array with class labels}
}}
\usage{
chess
}
\description{
Bidimensional binary dataset to be used as a toy example.
}
\keyword{datasets}

