% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covertype}
\alias{covertype}
\title{Covertype Data}
\format{A data.frame with 50000 rows and 55 variables:
\describe{
  \item{V1...V54}{binary features}
  \item{Target}{class labels}
}}
\usage{
covertype
}
\description{
Sample of size 50k from the multiclass dataset Covertype from the UCI repository.
}
\keyword{datasets}

