% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{fastknnCV}
\alias{fastknnCV}
\title{Cross-Validation for fastknn}
\usage{
fastknnCV(x, y, k = 3:15, method = "dist", folds = 5,
  eval.metric = "overall_error")
}
\arguments{
\item{x}{input matrix of dimension \code{nobs x nvars}.}

\item{y}{factor array wtih class labels for the \code{x} rows.}

\item{k}{sequence of possible k values to be evaluated (default is [3:15]).}

\item{method}{the probability estimator as in \code{fastknn}.}

\item{folds}{number of folds (default is 5) or an array with fold ids between 
1 and \code{n} identifying what fold each observation is in. The fold 
assigment given by \code{fastknnCV} does stratified sampling.}

\item{eval.metric}{loss to use for cross-validation. Currently five options are available:
\itemize{
 \item \code{eval.metric="overal_error"}: default option. It gives the overall 
 misclassification rate.
 \item \code{eval.metric="mean_error"}: gives the average in-class 
 misclassification rate.
 \item \code{eval.metric="auc"}: gives the average in-class area under the ROC curve.
 \item \code{eval.metric="logloss"}: gives the cross-entropy or logarithmic loss. 
}}
}
\value{
\code{list} with cross-validation results:
\itemize{
 \item \code{best_eval}: the best loss measure found in the 
 cross-validation procedure.
 \item \code{best_k}: the best k value found in the cross-validation procedure.
 \item \code{cv_table}: \code{data.frame} with the test performances for each k 
 on each data fold. 
}
}
\description{
Does n-fold cross-validation for \code{fastknn} to find the best k parameter.
}

