% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R
\name{knnDecision}
\alias{knnDecision}
\title{Classification Decision Boundary}
\usage{
knnDecision(xtr, ytr, xte, yte, k, method = "dist", dpi = 150)
}
\arguments{
\item{xtr}{matrix containing the training instances. If \code{xtr} is not a 
bidimensional matrix only the first two columns will be considered.}

\item{ytr}{factor array with the training labels.}

\item{xte}{(optional) Matrix containing the test instances. The test points will be 
plotted over the surface boundary. If missing, the training points will be 
plotted instead. If \code{xte} is not a bidimensional matrix only the first 
two columns will be considered.}

\item{yte}{(optional) Factor array with the test labels.}

\item{k}{number of neighbors considered.}

\item{method}{method used to infer the class membership probabilities of the 
test instances. See \code{\link{fastknn}} for more details.}

\item{dpi}{a scalar that defines the graph resolution (default = 150). 
It means that \code{dpi^2} data points will be generated from the original 
dataset to draw the decision boundary. So, for large values (>= 300) it may 
take too much time to plot.}
}
\value{
\code{ggplot2} object.
}
\description{
Decision boundary of the \code{fastknn} classifier.
}
\examples{
\dontrun{
library("caTools")
library("fastknn")

data("spirals")

x <- data.matrix(spirals$x)
y <- spirals$y

set.seed(2048)
tr.idx <- which(sample.split(Y = y, SplitRatio = 0.7))
x.tr <- x[tr.idx,]
x.te <- x[-tr.idx,]
y.tr <- y[tr.idx]
y.te <- y[-tr.idx]

knnDecision(xtr = x.tr, ytr = y.tr, xte = x.te, yte = y.te, k = 10)
}
}
\author{
David Pinto.
}

