% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multi_spirals}
\alias{multi_spirals}
\title{Multi Spirals Dataset}
\format{A data.frame with 2000 rows and 3 variables:
\describe{
  \item{x1}{random generated}
  \item{x2}{random generated}
  \item{target}{class labels}
}}
\usage{
multi_spirals
}
\description{
Bidimensional multiclass dataset to be used in toy examples.
}
\keyword{datasets}

